refactor: turning developer menu into a service.
