+++ TaskManager.DomainLayer/Model/Developer.cs
@@ -1,50 +1,20 @@
-﻿using TaskManager.ConsoleInteraction;
-using TaskManager.ConsoleInteraction.Components;
+﻿
+using TaskManager.DomainLayer.Service;
 
 namespace TaskManager.DomainLayer.Model
 {
     internal class Developer : User
     {
+        private readonly DeveloperMenuService _developerMenuService;
         public Developer(string newName, string newLogin, string? newEmail = null) : base(newName, newLogin, newEmail)
         {
             SetJob(JobEnum.Developer);
+            _developerMenuService = new DeveloperMenuService(this);
         }
 
         public override void Greeting()
         {
-            MainMenu();
-        }
-
-        public void MainMenu()
-        {
-            string[] menuDev = { "Alterar senha", "Sair" };
-            Menu options = new Menu(menuDev);
-
-            while (true)
-            {
-                Console.Clear();
-                int selectedOption = options.DisplayMenu(Title.GreetingDev());
-                if (!AnalyzeMainMenu(selectedOption))
-                {
-                    break;
-                }
-            }
-        }
-
-        private bool AnalyzeMainMenu(int selectedOption)
-        {
-            switch (selectedOption)
-            {
-                case 0:
-                    base.TryChangingPassword();
-                    return true;
-                case 1:
-                    Message.Returning();
-                    return false;
-                default:
-                    Console.WriteLine("Opção inválida");
-                    return true;
-            }
+            _developerMenuService.ShowMainMenu();
         }
     }
 }
